Use `AssemblyQualifiedName` instead of `ToString()` for the string representation of generic type arguments. Without assembly information loading the type can fail on the receiving side.
Also, set `throwOnError` to `true` when loading the type on the receiving side, failing immediately if the type cannot be loaded. Without that, the generic argument is null which leads to a `NullReferenceException` much later (when looking up the type in the cache, which assumes keys are non-null), making it difficult to see the root cause.